change module name to ci_provider
