+++ pkg/identity/ci_provider/principal.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package generic
+package ci_provider
 
 import (
 	"context"
